+++ RollbarDeploys/Tests/RollbarDeploysTests-ObjC/RollbarDeploysTests.m
@@ -1,6 +1,7 @@
 //  Copyright (c) 2018 Rollbar, Inc. All rights reserved.
 
 #import <XCTest/XCTest.h>
+#import "../../../UnitTests/RollbarUnitTestSettings.h"
 
 @import RollbarDeploys;
 
@@ -48,7 +49,7 @@
 @implementation RollbarDeploysTests
 
 - (void)testDeploymentDto {
-    NSString * const environment = @"Rollbar-Apple-UnitTests";
+    NSString * const environment = ROLLBAR_UNIT_TEST_ENVIRONMENT;
     NSString * const comment = @"a new deploy";
     NSString * const revision = @"a_revision";
     NSString * const localUsername = @"UnitTestRunner";
@@ -77,7 +78,7 @@
     NSDateFormatter *dateFormatter=[[NSDateFormatter alloc] init];
     [dateFormatter setDateFormat:@"yyyy-MM-dd HH:mm:ss"];
     
-    NSString * const environment = @"Rollbar-Apple-UnitTests";
+    NSString * const environment = ROLLBAR_UNIT_TEST_ENVIRONMENT;
     NSString * const comment =
         [NSString stringWithFormat:@"a new deploy at %@", [dateFormatter stringFromDate:[NSDate date]]];
     NSString * const revision = @"a_revision";
@@ -91,8 +92,8 @@
                                                                      localUserName:localUsername
                                                                    rollbarUserName:rollbarUsername];
     RollbarDeploysManager *deploysManager =
-        [[RollbarDeploysManager alloc] initWithWriteAccessToken:@"efdc4b85d66045f293a7f9e99c732f61"
-                                                readAccessToken:@"595cbf76b05b45f2b3ef661a2e0078d4"
+        [[RollbarDeploysManager alloc] initWithWriteAccessToken:ROLLBAR_UNIT_TEST_DEPLOYS_WRITE_ACCESS_TOKEN
+                                                readAccessToken:ROLLBAR_UNIT_TEST_DEPLOYS_READ_ACCESS_TOKEN
                                  deploymentRegistrationObserver:observer
                                       deploymentDetailsObserver:observer
                                   deploymentDetailsPageObserver:observer
@@ -104,8 +105,8 @@
     NSString * const testDeploymentId = @"23922850";
     RollbarDeploysObserver *observer = [[RollbarDeploysObserver alloc] init];
     RollbarDeploysManager *deploysManager =
-    [[RollbarDeploysManager alloc] initWithWriteAccessToken:@"efdc4b85d66045f293a7f9e99c732f61"
-                                            readAccessToken:@"595cbf76b05b45f2b3ef661a2e0078d4"
+    [[RollbarDeploysManager alloc] initWithWriteAccessToken:ROLLBAR_UNIT_TEST_DEPLOYS_WRITE_ACCESS_TOKEN
+                                            readAccessToken:ROLLBAR_UNIT_TEST_DEPLOYS_READ_ACCESS_TOKEN
                              deploymentRegistrationObserver:observer
                                   deploymentDetailsObserver:observer
                               deploymentDetailsPageObserver:observer
@@ -116,8 +117,8 @@
 - (void)testGetDeploymentsPage {
     RollbarDeploysObserver *observer = [[RollbarDeploysObserver alloc] init];
     RollbarDeploysManager *deploysManager =
-    [[RollbarDeploysManager alloc] initWithWriteAccessToken:@"efdc4b85d66045f293a7f9e99c732f61"
-                                            readAccessToken:@"595cbf76b05b45f2b3ef661a2e0078d4"
+    [[RollbarDeploysManager alloc] initWithWriteAccessToken:ROLLBAR_UNIT_TEST_DEPLOYS_WRITE_ACCESS_TOKEN
+                                            readAccessToken:ROLLBAR_UNIT_TEST_DEPLOYS_READ_ACCESS_TOKEN
                              deploymentRegistrationObserver:observer
                                   deploymentDetailsObserver:observer
                               deploymentDetailsPageObserver:observer