+++ RollbarNotifier/Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.2
+// swift-tools-version:5.5
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -22,6 +22,7 @@ let package = Package(
         // Dependencies declare other packages that this package depends on.
         // .package(url: /* package url */, from: "1.0.0"),
         .package(path: "../RollbarCommon"),
+        .package(path: "../UnitTesting"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -41,14 +42,9 @@ let package = Package(
         ),
         .testTarget(
             name: "RollbarNotifierTests",
-            dependencies: ["RollbarNotifier"],
-            cSettings: [
-                .headerSearchPath("Tests/RollbarNotifierTests/**"),
-//                .headerSearchPath("Sources/RollbarNotifier"),
-//                .headerSearchPath("Sources/RollbarNotifier/include"),
-//                .headerSearchPath("Sources/RollbarNotifier/DTOs"),
-                
-//                .define("DEFINES_MODULE"),
+            dependencies: [
+                "UnitTesting",
+                "RollbarNotifier"
             ]
         ),
         .testTarget(