+++ UnitTesting/Sources/UnitTesting/RollbarUnitTestSettings.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+public struct RollbarUnitTestSettings {
+    public static let environment              = "Rollbar-Apple-UnitTests";
+    public static let payloadsPostAccessToken  = "09da180aba21479e9ed3d91e0b8d58d6";
+    public static let deploysWriteAccessToken  = "efdc4b85d66045f293a7f9e99c732f61";
+    public static let deploysReadAccessToken   = "595cbf76b05b45f2b3ef661a2e0078d4";
+}