+++ UnitTests/RollbarUnitTestSettings.h
@@ -1,0 +1,16 @@
+#ifndef RollbarUnitTestSettings_h
+#define RollbarUnitTestSettings_h
+
+static NSString * const ROLLBAR_UNIT_TEST_ENVIRONMENT =
+@"Rollbar-Apple-UnitTests";
+
+static NSString * const ROLLBAR_UNIT_TEST_PAYLOADS_ACCESS_TOKEN =
+@"09da180aba21479e9ed3d91e0b8d58d6";
+
+static NSString * const ROLLBAR_UNIT_TEST_DEPLOYS_WRITE_ACCESS_TOKEN =
+@"efdc4b85d66045f293a7f9e99c732f61";
+
+static NSString * const ROLLBAR_UNIT_TEST_DEPLOYS_READ_ACCESS_TOKEN =
+@"595cbf76b05b45f2b3ef661a2e0078d4";
+
+#endif /* RollbarUnitTestSettings_h */